update crontab for celery

+++ update_valutes_celery/celery.py
@@ -15,15 +15,23 @@ celery_app.config_from_object('api.settings', namespace='CELERY')
 # celery_app.autodiscover_tasks(lambda: settings.INSTALLED_APPS)
 
 from valutes.models import Valute
+import datetime
 
-@celery_app.task
+@celery_app.task(name="update valutes")
 def update_valutes():
+    currentMinutes = datetime.datetime.now().strftime('%M')
+    if int(currentMinutes) not in [5, 10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 0]:
+        print('idle %s' % datetime.datetime.now())
+        return
+
     import requests
     VALUTE_URL = 'https://www.cbr-xml-daily.ru/daily_json.js'
     response = requests.get(url = VALUTE_URL)
     data = response.json()
+    print('delete valutes')
     Valute.objects.all().delete()
 
+    print('create valutes')
     for key, item in data['Valute'].items():
         Valute.objects.create(
             external_id=item["ID"],
@@ -38,10 +46,5 @@ def update_valutes():
 
 @celery_app.on_after_configure.connect
 def setup_periodic_tasks(sender, **kwargs):
-    sender.add_periodic_task(
-        crontab(minute='*/5'),
-        update_valutes.s(),
-        name="update valutes"
-    )
-#     sender.add_periodic_task(10.0, update_valutes.s(), name="update valutes")
+    sender.add_periodic_task(60.0, update_valutes.s(), name="update valutes")
